@@ -3,16 +3,16 @@ apply plugin: "kotlin-android"
 apply plugin: "kotlin-kapt"
 apply plugin: "kotlin-android-extensions"
 apply plugin: "androidx.navigation.safeargs"
-apply plugin: "io.fabric"
+apply plugin: "com.google.firebase.crashlytics"
 
 android {
-    compileSdkVersion 29
+    compileSdkVersion 30
     defaultConfig {
         applicationId "pl.kapucyni.wolczyn.app"
         minSdkVersion 21
-        targetSdkVersion 29
-        versionCode 11
-        versionName "3.0.0"
+        targetSdkVersion 30
+        versionCode 12
+        versionName "2021.0.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
