@@ -1,26 +1,28 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = "1.3.41"
+    ext.kotlin_version = "1.5.10"
     repositories {
-        mavenCentral()
         google()
+        mavenCentral()
+        //noinspection JcenterRepositoryObsolete
         jcenter()
 
         maven { url "https://maven.fabric.io/public" }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.4.2'
+        classpath 'com.android.tools.build:gradle:3.4.3'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath "android.arch.navigation:navigation-safe-args-gradle-plugin:1.0.0"
-        classpath "com.google.gms:google-services:4.2.0"
-        classpath "io.fabric.tools:gradle:1.29.0"
+        classpath "com.google.gms:google-services:4.3.8"
+        classpath "com.google.firebase:firebase-crashlytics-gradle:2.6.1"
     }
 }
 
 allprojects {
     repositories {
         google()
+        //noinspection JcenterRepositoryObsolete
         jcenter()
 
         maven { url "https://jitpack.io" }
