@@ -9,14 +9,15 @@ import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
+import android.view.WindowInsetsController
 import android.widget.Toast
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.ActionBarDrawerToggle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.core.view.GravityCompat
 import androidx.core.view.children
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
+import androidx.lifecycle.ViewModelProvider
 import com.google.android.material.navigation.NavigationView
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.app_bar_main.*
@@ -42,24 +43,30 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
 
     override fun onCreate(savedInstanceState: Bundle?) {
         setTheme(R.style.AppTheme)
-        if (PreferencesManager.getNightMode()) {
+        if (PreferencesManager.getNightMode())
             AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
-            window.decorView.systemUiVisibility = 0
-        } else {
+        else
             AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-        }
 
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         setSupportActionBar(toolbar)
         supportActionBar?.setDisplayShowTitleEnabled(false)
 
-        if (!PreferencesManager.getNightMode() && Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR
-            window.statusBarColor = Color.WHITE
+        if (!PreferencesManager.getNightMode()) {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R)
+                window.decorView.windowInsetsController?.setSystemBarsAppearance(
+                    WindowInsetsController.APPEARANCE_LIGHT_STATUS_BARS,
+                    WindowInsetsController.APPEARANCE_LIGHT_STATUS_BARS
+                )
+            else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                @Suppress("DEPRECATION")
+                window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR
+                window.statusBarColor = Color.WHITE
+            }
         }
 
-        mViewModel = ViewModelProviders.of(this@MainActivity).get(MainViewModel::class.java)
+        mViewModel = ViewModelProvider(this@MainActivity).get(MainViewModel::class.java)
 
         val toggle = ActionBarDrawerToggle(
             this@MainActivity, drawerLayout, toolbar,
@@ -70,10 +77,17 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
         toggle.syncState()
 
         navView.setNavigationItemSelectedListener(this)
-        onNavigationItemSelected(savedInstanceState?.let { navView.menu.findItem(it.getInt("selectedItem", 0)) }
+        onNavigationItemSelected(savedInstanceState?.let {
+            navView.menu.findItem(
+                it.getInt(
+                    "selectedItem",
+                    0
+                )
+            )
+        }
             ?: navView.menu.getItem(0))
 
-        mViewModel.currentUser.observe(this@MainActivity, Observer { user ->
+        mViewModel.currentUser.observe(this@MainActivity, { user ->
             if (user != null && headerDrawerImage != null) {
                 GlideApp.with(this@MainActivity)
                     .load(user.photo_url)
@@ -120,16 +134,6 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
         if (mIsTransactionPending) goToSelectedFragment(mCurrentFragmentId!!)
     }
 
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if (requestCode == 12 && resultCode == Activity.RESULT_OK) {
-            if (!data!!.getBooleanExtra("loginSuccess", false) && data.getIntExtra("fragment", 0) == R.id.nav_group)
-                goBackToHome()
-            else
-                goToSelectedFragment(data.getIntExtra("fragment", 0))
-        }
-        super.onActivityResult(requestCode, resultCode, data)
-    }
-
     override fun onBackPressed() {
         when {
             drawerLayout.isDrawerOpen(GravityCompat.START) -> drawerLayout.closeDrawer(GravityCompat.START)
@@ -148,7 +152,8 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
 
     private fun doubleBackPressToExit() {
         if (mBackPressed + 2000 > System.currentTimeMillis()) super.onBackPressed()
-        else Toast.makeText(baseContext, getString(R.string.press_to_exit), Toast.LENGTH_SHORT).show()
+        else Toast.makeText(baseContext, getString(R.string.press_to_exit), Toast.LENGTH_SHORT)
+            .show()
         mBackPressed = System.currentTimeMillis()
     }
 
@@ -183,7 +188,8 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
             }
             R.id.action_settings -> {
                 toolbar.title = getString(R.string.preferences)
-                supportFragmentManager.beginTransaction().replace(R.id.container, PreferenceFragment()).commit()
+                supportFragmentManager.beginTransaction()
+                    .replace(R.id.container, PreferenceFragment()).commit()
                 mCurrentFragmentId = -1 * navView.menu.children.indexOfFirst { it.isChecked } - 1
                 true
             }
@@ -256,7 +262,8 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
         }
 
         fragment?.let {
-            if (mIsTransactionSafe) supportFragmentManager.beginTransaction().replace(R.id.container, it).commit()
+            if (mIsTransactionSafe) supportFragmentManager.beginTransaction()
+                .replace(R.id.container, it).commit()
             else mIsTransactionPending = true
         }
         return true
@@ -270,16 +277,24 @@ class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelecte
     }
 
     private fun openLoginActivity() {
-        val intent =
-            Intent(this@MainActivity, LoginActivity::class.java).putExtra("fragment", mCurrentFragmentId)
-        startActivityForResult(intent, 12)
+        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
+            if (it.resultCode == Activity.RESULT_OK) {
+                if (!it.data!!.getBooleanExtra("loginSuccess", false)
+                    && it.data!!.getIntExtra("fragment", 0) == R.id.nav_group
+                ) goBackToHome()
+                else goToSelectedFragment(it.data!!.getIntExtra("fragment", 0))
+            }
+        }.launch(
+            Intent(this@MainActivity, LoginActivity::class.java)
+                .putExtra("fragment", mCurrentFragmentId)
+        )
     }
 
     fun openArchiveDetailsActivity(meeting: ArchiveMeeting) {
         val intent = Intent(this@MainActivity, ArchiveMeetingDetailsActivity::class.java)
         intent.putExtra("title", meeting.name.split(" - ")[1])
         intent.putParcelableArrayListExtra("records", meeting.records)
-        startActivityForResult(intent, 21)
+        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {}.launch(intent)
     }
 
     fun changeToolbarTitle(newTitle: String) {
