@@ -13,13 +13,13 @@ class KapucyniApiRepository(private val api: KapucyniApi) {
     suspend fun loginToSystemWithEmail(login: String, password: String): String? {
         val response = api.loginToSystemWithEmailAsync(login, password, BuildConfig.KAPUCYNI_API_KEY)
         if (response.isSuccessful) return response.body()
-        else throw IllegalStateException(response.errorBody()?.string())
+        else throw IllegalStateException(response.errorBody()?.toString())
     }
 
     suspend fun loginToSystemWithSocial(email: String, identifier: String, media: String): String? {
         val response = api.loginToSystemWithSocialAsync(email, identifier, media, BuildConfig.KAPUCYNI_API_KEY)
         if (response.isSuccessful) return response.body()
-        else throw IllegalStateException(response.errorBody()?.string())
+        else throw IllegalStateException(response.errorBody()?.toString())
     }
 
     suspend fun getUserInfo(): Response<User> = api.getUserInfoAsync()
