@@ -1,11 +1,10 @@
 package pl.kapucyni.wolczyn.app.utils
 
 import android.content.Context
-import com.bumptech.glide.annotation.GlideModule
-import com.bumptech.glide.module.AppGlideModule
 import com.bumptech.glide.Glide
 import com.bumptech.glide.Registry
-import java.io.InputStream
+import com.bumptech.glide.annotation.GlideModule
+import com.bumptech.glide.module.AppGlideModule
 
 @GlideModule
 class AppGlideModule : AppGlideModule() {
