@@ -8,8 +8,9 @@ import kotlinx.android.synthetic.main.item_archive_record.view.*
 import pl.kapucyni.wolczyn.app.R
 import pl.kapucyni.wolczyn.app.model.Record
 import pl.kapucyni.wolczyn.app.utils.openWebsiteInCustomTabsService
+import java.util.*
 
-class ArchiveMeetingsRecyclerAdapter(private val mRecords: ArrayList<Record>) :
+class ArchiveMeetingsRecyclerAdapter(private val mRecords: ArrayList<Record>?) :
     RecyclerView.Adapter<ArchiveMeetingsRecyclerAdapter.ArchiveMeetingViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ArchiveMeetingViewHolder =
@@ -18,9 +19,9 @@ class ArchiveMeetingsRecyclerAdapter(private val mRecords: ArrayList<Record>) :
         )
 
     override fun onBindViewHolder(holder: ArchiveMeetingViewHolder, position: Int) =
-        holder.bindView(mRecords[position])
+        holder.bindView(mRecords?.get(position) ?: Record())
 
-    override fun getItemCount(): Int = mRecords.size
+    override fun getItemCount(): Int = mRecords?.size ?: 0
 
     inner class ArchiveMeetingViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         fun bindView(record: Record) {
