@@ -5,20 +5,20 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
+import androidx.preference.SwitchPreference
 import pl.kapucyni.wolczyn.app.R
 import pl.kapucyni.wolczyn.app.utils.PreferencesManager
 
 class PreferenceFragment : PreferenceFragmentCompat() {
 
-
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         addPreferencesFromResource(R.xml.preferences)
-
-        preferenceManager.findPreference(getString(R.string.night_mode_key)).onPreferenceChangeListener =
+        preferenceManager.findPreference<SwitchPreference>(getString(R.string.night_mode_key))?.onPreferenceChangeListener =
             Preference.OnPreferenceChangeListener { _, newValue ->
                 activity?.let {
                     AppCompatDelegate.setDefaultNightMode(
-                        if (newValue as Boolean) AppCompatDelegate.MODE_NIGHT_YES else AppCompatDelegate.MODE_NIGHT_NO
+                        if (newValue as Boolean) AppCompatDelegate.MODE_NIGHT_YES
+                        else AppCompatDelegate.MODE_NIGHT_NO
                     )
                     PreferencesManager.setNightMode(newValue)
                 }
