@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProviders
+import androidx.lifecycle.ViewModelProvider
 import kotlinx.android.synthetic.main.fragment_signings.view.*
 import pl.kapucyni.wolczyn.app.R
 import pl.kapucyni.wolczyn.app.utils.tryToRunFunctionOnInternet
@@ -17,17 +17,20 @@ class SigningsFragment : Fragment() {
 
     private lateinit var mViewModel: MainViewModel
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
-        inflater.inflate(R.layout.fragment_signings, container, false)
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
+    ): View? = inflater.inflate(R.layout.fragment_signings, container, false)
 
     @SuppressLint("SetJavaScriptEnabled")
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         view.loadingIndicator.hide()
-        activity?.let { mViewModel = ViewModelProviders.of(it).get(MainViewModel::class.java) }
+        activity?.let { mViewModel = ViewModelProvider(it).get(MainViewModel::class.java) }
         view.webView.settings.javaScriptEnabled = true
         if (savedInstanceState != null) view.webView.restoreState(savedInstanceState)
-        else activity?.tryToRunFunctionOnInternet { mViewModel.loadSignings(view, (activity as MainActivity)) }
+        else activity?.tryToRunFunctionOnInternet {
+            mViewModel.loadSignings(view, (activity as MainActivity))
+        }
     }
 
     override fun onSaveInstanceState(outState: Bundle) {
