@@ -6,8 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
+import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.DefaultItemAnimator
 import androidx.recyclerview.widget.LinearLayoutManager
 import kotlinx.android.synthetic.main.fragment_weather.view.*
@@ -24,8 +23,9 @@ class WeatherFragment : Fragment() {
     private lateinit var mMainViewModel: MainViewModel
     private lateinit var mAdapter: WeatherRecyclerAdapter
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
-        inflater.inflate(R.layout.fragment_weather, container, false)
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
+    ): View? = inflater.inflate(R.layout.fragment_weather, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -35,21 +35,26 @@ class WeatherFragment : Fragment() {
         view.weatherRecyclerView.itemAnimator = DefaultItemAnimator()
         view.weatherRecyclerView.adapter = mAdapter
 
-        mWeatherViewModel = ViewModelProviders.of(this@WeatherFragment).get(WeatherViewModel::class.java)
+        mWeatherViewModel =
+            ViewModelProvider(this@WeatherFragment).get(WeatherViewModel::class.java)
         activity?.let {
-            mMainViewModel = ViewModelProviders.of(it).get(MainViewModel::class.java)
+            mMainViewModel = ViewModelProvider(it).get(MainViewModel::class.java)
             if (mMainViewModel.weatherList == null) fetchWeather()
             else mWeatherViewModel.weatherRecords.postValue(mMainViewModel.weatherList)
         }
 
-        mWeatherViewModel.weatherRecords.observe(this@WeatherFragment, Observer { weatherList ->
+        mWeatherViewModel.weatherRecords.observe(viewLifecycleOwner, { weatherList ->
             val days = weatherList.map { it.dt_txt.split(" ")[0] }.distinct()
             var weatherDays = arrayListOf<List<WeatherRecord>>()
             days.forEach { day ->
                 weatherDays.add(weatherList.filter {
                     val dayOfMonth = day.split("-").last().toInt() + 1
                     val nextDayMidnight =
-                        day.replaceRange(8, 10, if (dayOfMonth < 10) "0$dayOfMonth" else "$dayOfMonth") + " 00:00:00"
+                        day.replaceRange(
+                            8,
+                            10,
+                            if (dayOfMonth < 10) "0$dayOfMonth" else "$dayOfMonth"
+                        ) + " 00:00:00"
                     it.dt_txt.split(" ")[0] == day || it.dt_txt == nextDayMidnight
                 })
             }
